Flux balance function (#14)
* Add ideal SERMeQ time evolution

Force a SERMeQ initial profile and time series evolution with flux from upstream.  Plot to compare with what OGGM k-calving is doing.

* Add flux_evolve function

Add function to step in time with variable influx, and add necessary attributes of `PlasticGlacier` to support this.

`flux_evolve` is not matching the profiles produced by the for-loop in `idealized_feedbacks.ipynb` for whatever reason. Debugging still needed.

* Implement PlasticGlacier.flux_evolve

Add `flux_evolve` function to chakra_sermeq.  Replace loop with function.

Co-authored-by: Fabien Maussion <fabien.maussion@uibk.ac.at>